+++ wonpick/src/main/java/com/won/wonpick/post/controller/PostController.java
@@ -142,4 +142,5 @@ public class PostController {
 		return "redirect:/post/searchPage";
 	}
 	
+	
 }
